Unify expression level and feature paging logic
Also create backend_iterators file

Issue #1371
