+++ tests/unit/test_imports.py
@@ -154,7 +154,8 @@ class ImportGraphLayerChecker(object):
         'cli': ['ga4gh/cli.py'],
         'client': ['ga4gh/client.py'],
         'frontend': ['ga4gh/frontend.py', 'ga4gh/repo_manager.py'],
-        'backend': ['ga4gh/backend.py', 'ga4gh/datarepo.py'],
+        'backend': ['ga4gh/backend.py', 'ga4gh/datarepo.py',
+                    'ga4gh/paging.py'],
         'exceptions': ['ga4gh/exceptions.py'],
         'datamodel': ['ga4gh/datamodel/bio_metadata.py',
                       'ga4gh/datamodel/reads.py',
